% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_A}
\alias{calc_A}
\title{Documentation Still Needed}
\usage{
calc_A(x_obs, mu, sigma)
}
\arguments{
\item{df}{A numeric matrix.}

\item{tnew}{A numeric matrix.}

\item{nearest_x}{A numeric vector.}
}
\value{
A numeric matrix.
}
\description{
This is a function that still needs to be documented properly.
}
