% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pme.R
\name{calc_coefficients}
\alias{calc_coefficients}
\title{Compute Spline Coefficients}
\usage{
calc_coefficients(X, t, W, w)
}
\arguments{
\item{X}{Numeric matrix of high-dimensional data.}

\item{t}{Numeric matrix of low-dimensional paramterizations.}

\item{W}{Numeric matrix of cluster weights.}

\item{w}{smoothing parameter.}
}
\description{
Compute Spline Coefficients
}
