% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_init_param}
\alias{calc_init_param}
\title{Documentation Still Needed}
\usage{
calc_init_param(df, tnew, nearest_x)
}
\arguments{
\item{df}{A numeric matrix.}

\item{tnew}{A numeric matrix.}

\item{nearest_x}{A numeric vector.}
}
\value{
A numeric matrix.
}
\description{
This is a function that still needs to be documented properly.
}
