% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_nearest_x}
\alias{calc_nearest_x}
\title{Documentation Still Needed}
\usage{
calc_nearest_x(df, x)
}
\arguments{
\item{df}{A numeric matrix.}

\item{x}{A numeric matrix.}
}
\value{
A numeric vector.
}
\description{
This is a function that still needs to be documented properly.
}
