% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_weights_cpp}
\alias{calc_weights_cpp}
\title{Calculate Weights of Mixture Components}
\usage{
calc_weights_cpp(x_obs, mu, sigma, epsilon, max_iter)
}
\arguments{
\item{x_obs}{A numeric matrix containing the unreduced data.}

\item{mu}{A numeric matrix of component centers.}

\item{sigma}{A numeric value denoting the bandwidth of the density estimation.}

\item{epsilon}{A numeric value denoting the tolerance of the Euclidean distance between weights.}

\item{max_iter}{An integer denoting the maximum number of iterations.}
}
\value{
A numeric vector of weights.
}
\description{
Calculate Weights of Mixture Components
}
