% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcE}
\alias{calcE}
\title{Documentation Still Needed}
\usage{
calcE(x, lambda)
}
\arguments{
\item{x}{Numeric matrix of values.}

\item{lambda}{Number of dimensions.}
}
\value{
A numeric matrix.
}
\description{
This is a function that still needs to be documented properly.
}
