% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcE_sphere}
\alias{calcE_sphere}
\title{Compute E matrix using spherical kernel}
\usage{
calcE_sphere(x, m = 2L)
}
\arguments{
\item{x}{Numeric matrix of values.}

\item{m}{Spline order.}
}
\value{
A numeric matrix.
}
\description{
Compute E matrix using spherical kernel
}
