% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_euclidean}
\alias{dist_euclidean}
\title{Compute the Euclidean Distance between Vectors}
\usage{
dist_euclidean(x, y)
}
\arguments{
\item{x, }{y Numeric vectors of values.}
}
\value{
A numeric value.
}
\description{
C++ implementation of Euclidean distance calculations for use in
internal Rcpp functions.
}
\examples{
x <- 1:10
y <- 91:100
dist_euclidean(x, y)
}
