% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eta_kernel}
\alias{eta_kernel}
\title{Documentation Still Needed}
\usage{
eta_kernel(t, lambda)
}
\arguments{
\item{t}{Numeric vector of values.}

\item{lambda}{Number of dimensions.}
}
\value{
A numeric value.
}
\description{
This is a function that still needs to be documented properly.
}
