% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{etaFunc}
\alias{etaFunc}
\title{Documentation Still Needed}
\usage{
etaFunc(t, tau, lambda)
}
\arguments{
\item{t}{A numeric vector of values.}

\item{tau}{A numeric matrix of values.}

\item{lambda}{The number of dimensions.}
}
\value{
A numeric matrix.
}
\description{
This is a function that still needs to be documented properly.
}
