% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fibonacci_sphere}
\alias{fibonacci_sphere}
\title{Sample points from the Fibonacci Sphere}
\usage{
fibonacci_sphere(n)
}
\arguments{
\item{n}{The number of points}
}
\value{
A matrix of points, using spherical coordinates.
}
\description{
Sample points from the Fibonacci Sphere
}
