% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdmde.R
\name{hdmde_mod}
\alias{hdmde_mod}
\title{High-Dimensional Mixture Density Estimation - Modification}
\usage{
hdmde_mod(x_obs, N0, alpha, max_comp)
}
\arguments{
\item{x_obs}{A numeric matrix containing the data to be reduced.}

\item{N0}{An integer specifying the lower bound for the number of density components.}

\item{alpha}{A numeric value between 0 and 1 specifying the confidence level.}

\item{max_comp}{An integer specifying the upper bound for the number of density components.}
}
\value{
An object of type hdmde
}
\description{
Data reduction approach to estimate a high-dimensional dataset using a
mixture of distributions. The centers of these distributions are identified
using k-means clustering. This function is modified to use the BIC of the
k-means clustering output to select the number of mixture components.
}
