% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pme.R
\name{initialize_pme}
\alias{initialize_pme}
\title{Create PME Initialization}
\usage{
initialize_pme(
  x,
  d,
  min_clusters,
  alpha,
  max_clusters,
  component_type = "centers",
  algorithm = "isomap",
  subsample_size = 5,
  template = "euclidean"
)
}
\arguments{
\item{x}{A numeric matrix of data.}

\item{d}{The intrinsic dimension.}

\item{min_clusters}{The minimum number of mixture components.}

\item{alpha}{Significance level.}

\item{max_clusters}{Maximum number of components.}

\item{component_type}{Whether to represent clusters using centers or subsampled points: centers or subsample.}

\item{algorithm}{The manifold learning algorithm to use for initial parameterization.}

\item{subsample_size}{If component_type is "subsample", the number of points to include.}

\item{template}{The template manifold to use.}
}
\value{
A list used to initialize PME.
}
\description{
Create PME Initialization
}
