% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpme.R
\name{is_lpme}
\alias{is_lpme}
\title{Check Whether Object is of Type LPME}
\usage{
is_lpme(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
A logical value.
}
\description{
Check Whether Object is of Type LPME
}
