% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_pme.R
\name{is_pme}
\alias{is_pme}
\title{Is Object PME}
\usage{
is_pme(x)
}
\arguments{
\item{x}{An object.}
}
\value{
Logical value.
}
\description{
Check Whether Object Has Type PME
}
\examples{
num_value <- 5
is_pme(num_value)
}
