% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_smoothing_kernel}
\alias{log_smoothing_kernel}
\title{Smoothing Kernel for Density Estimation}
\usage{
log_smoothing_kernel(x, mu, sigma)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{mu}{The mean of the Gaussian density}

\item{sigma}{The standard deviation of the Gaussian density}
}
\value{
A numeric value.
}
\description{
Implements Gaussian kernel smoothing on log-scale
}
