% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logspace_diff}
\alias{logspace_diff}
\title{Difference in Logspace}
\usage{
logspace_diff(lx, ly)
}
\arguments{
\item{lx}{A double value in logspace}

\item{ly}{A double value in logspace}
}
\value{
A numeric value
}
\description{
Given log(x) and log(y), return log(x-y)
}
