% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logspace_sum_vec}
\alias{logspace_sum_vec}
\title{Sum over a Vector in Logspace}
\usage{
logspace_sum_vec(x)
}
\arguments{
\item{x}{A numeric vector of values in logspace}
}
\value{
A numeric value
}
\description{
Recursively call logspace_sum() over n > 2 values
}
