% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pme.R
\name{pme}
\alias{pme}
\title{Principal Manifold Estimation}
\usage{
pme(
  data,
  d,
  initialization = NULL,
  lambda = exp(-15:5),
  alpha = 0.05,
  min_clusters = 0,
  max_clusters = 100,
  epsilon = 0.05,
  max_iter = 100,
  SSD_ratio_threshold = 10,
  print_plots = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data}{A numeric matrix of high-dimensional data.}

\item{d}{A positive integer representing the intrinsic dimension.}

\item{initialization}{A list of values providing an initialization for \code{pme()}. It is not recommended to supply these values manually.}

\item{lambda}{A vector of smoothing values to be considered.}

\item{alpha}{The significant level to be used when testing the need for additional clusters in data reduction.}

\item{min_clusters}{The minimum number of clusters allowed in data reduction.}

\item{max_clusters}{The maximum number of clusters allowed in data reduction.}

\item{epsilon}{Threshold for change in SSD to stop iterations for a given smoothing value.}

\item{max_iter}{The maximum number of iterations allowed for a given smoothing value.}

\item{SSD_ratio_threshold}{The maximum increase in SSD allowed before stopping iterations for a given smoothing value.}

\item{print_plots}{A logical value indicating whether plots of the estimated manifolds should be produced.}

\item{verbose}{A logical value indicating whether updates should be printed.}
}
\value{
An object of type pme.
}
\description{
This function still requires completed documentation
}
\examples{
r <- runif(1000, min = 0, max = pi)
x <- sin(r) + rnorm(1000, mean = 0, sd = 0.25)
mat <- cbind(r, x)
out <- pme(mat, d = 1)
}
