% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{projection_lpme}
\alias{projection_lpme}
\title{Project onto Low-Dimensional Manifold}
\usage{
projection_lpme(x, f, initial_guess, n_knots, d_new, gamma)
}
\arguments{
\item{x}{A value}

\item{f}{A value}

\item{initial_guess}{A value}

\item{n_knots}{A value}

\item{d_new}{A value}

\item{gamma}{A value}
}
\value{
A value
}
\description{
Project onto Low-Dimensional Manifold
}
