% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpparma_hello_world}
\alias{rcpparma_hello_world}
\title{RcppArmadillo Hello World}
\usage{
rcpparma_hello_world()
}
\description{
This is a simple example of creating two matrices and returning
the result of an operation on them.
}
\examples{
rcpparma_hello_world
}
