% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{smoothing_kernel}
\alias{smoothing_kernel}
\title{Smoothing Kernel for Density Estimation}
\usage{
smoothing_kernel(x, mu, sigma)
}
\arguments{
\item{x}{A vector of numeric values.}

\item{mu}{The mean of the Gaussian density.}

\item{sigma}{The standard deviation of the Gaussian density.}
}
\value{
A vector of numeric values.
}
\description{
Implements Gaussian kernel smoothing.
}
