% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{solve_spline}
\alias{solve_spline}
\title{Find the Coefficients of a Weighted Spline Function}
\usage{
solve_spline(E, t_val, X, w, d, D)
}
\arguments{
\item{E}{A numeric matrix.}

\item{t_val}{A numeric matrix.}

\item{X}{A numeric matrix.}

\item{w}{The smoothing parameter.}

\item{d}{The intrinsic dimension. (Note: Unused in function body)}

\item{D}{The dimension of the higher dimensional space. (Note: Unused in function body)}

\item{W}{A numeric matrix.}
}
\value{
A numeric matrix.
}
\description{
Find the Coefficients of a Weighted Spline Function
}
