% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{solve_weighted_spline_r}
\alias{solve_weighted_spline_r}
\title{Find the Coefficients of a Weighted Spline Function}
\usage{
solve_weighted_spline_r(E, W, t_val, X, w, d, D)
}
\arguments{
\item{E}{A numeric matrix.}

\item{W}{A numeric matrix.}

\item{t_val}{A numeric matrix.}

\item{X}{A numeric matrix.}

\item{w}{The smoothing parameter.}

\item{d}{The intrinsic dimension.}

\item{D}{The dimension of the higher dimensional space.}
}
\value{
A numeric matrix.
}
\description{
Find the Coefficients of a Weighted Spline Function
}
