% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sphere_kernel}
\alias{sphere_kernel}
\title{Compute spherical kernel}
\usage{
sphere_kernel(x, y, m)
}
\arguments{
\item{x}{Vector 1, expressed in cartesian coordinates}

\item{y}{Vector 2, expressed in cartesian coordinates}

\item{m}{Order of the spline function}
}
\value{
A numeric value.
}
\description{
Compute spherical kernel
}
