% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sphere_kernel_func}
\alias{sphere_kernel_func}
\title{Compute spherical kernel values for single observation}
\usage{
sphere_kernel_func(t, tau, m = 2L)
}
\arguments{
\item{t}{A numeric vector of values.}

\item{tau}{A numeric matrix of values.}

\item{m}{Spline order.}
}
\value{
A numeric matrix.
}
\description{
Compute spherical kernel values for single observation
}
