% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sphere_q}
\alias{sphere_q}
\title{Compute the Quadratic Functional for Spherical Kernel}
\usage{
sphere_q(z, m)
}
\arguments{
\item{z}{Cosine of angle between two vectors}

\item{m}{Order of the spline function}
}
\value{
A numeric value.
}
\description{
Closed-form solutions are given in Wahba (1981) Table 1
}
