% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_A}
\alias{calc_A}
\title{Calculate "A" Matrix}
\usage{
calc_A(x_obs, mu, sigma)
}
\arguments{
\item{x_obs}{A numeric matrix containing unreduced data.}

\item{mu}{A numeric matrix of component centers.}

\item{sigma}{A numeric value denoting the bandwidth of density estimation.}
}
\value{
A numeric matrix.
}
\description{
This is a function that still needs to be documented properly.
}
