% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpme.R
\name{lpme}
\alias{lpme}
\title{Fit an LPME Object}
\usage{
lpme(
  data,
  d,
  smoothing_method = "spline",
  gamma = NULL,
  lambda = NULL,
  initialization_algorithm = "isomap",
  init_type = "centers",
  alpha = 0.05,
  min_clusters = 0,
  max_clusters = 500,
  epsilon = 0.05,
  max_iter = 100,
  verbose = TRUE,
  print_plots = TRUE,
  increase_threshold = 1.05,
  init = "full"
)
}
\arguments{
\item{data}{A numeric matrix of data.}

\item{d}{The intrinsic dimension of the data.}

\item{smoothing_method}{The approach taken to smoothing over PME coefficients.}

\item{gamma}{A vector of numeric smoothing parameter values.}

\item{lambda}{A vector of numeric smoothing parameter values for the PME algorithm.}

\item{initialization_algorithm}{A character value describing the algorithm used to obtain initial parameter values. Accepted values are "isomap", "diffusion_maps", and "laplacian_eigenmaps".}

\item{init_type}{A character value describing whether initialization should use cluster centers or subsample from cluster values. Accepted values are "centers" or "subsampling".}

\item{alpha}{A value.}

\item{min_clusters}{The minimum number of clusters in the data.}

\item{max_clusters}{The maximum number of clusters identified in the data.}

\item{epsilon}{A value.}

\item{max_iter}{The maximum number of iterations.}

\item{verbose}{A logical value indicating whether messages should be printed.}

\item{print_plots}{A logical value indicating whether plots should be printed.}

\item{increase_threshold}{A value.}

\item{init}{Indicates which time points are used to initialize the function.}
}
\value{
An object of type "lpme".
}
\description{
Fit an LPME Object
}
