% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpme.R
\name{lpme}
\alias{lpme}
\title{Fit an LPME Object}
\usage{
lpme(
  df,
  d,
  tuning_para_seq = c(0, exp(-15:10)),
  alpha = 0.05,
  max_comp = 500,
  epsilon = 0.05,
  max_iter = 100,
  verbose = TRUE,
  print_plots = TRUE,
  increase_threshold = 1.05,
  init = "full"
)
}
\arguments{
\item{df}{A numeric matrix of data.}

\item{d}{The intrinsic dimension of the data.}

\item{tuning_para_seq}{A vector of numeric smoothing parameter values.}

\item{alpha}{A value.}

\item{max_comp}{The maximum number of clusters identified in the data.}

\item{epsilon}{A value.}

\item{max_iter}{The maximum number of iterations.}

\item{verbose}{A logical value indicating whether messages should be printed.}

\item{print_plots}{A logical value indicating whether plots should be printed.}

\item{increase_threshold}{A value.}

\item{init}{Indicates which time points are used to initialize the function.}
}
\value{
An object of type "lpme".
}
\description{
Fit an LPME Object
}
