% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpme.R
\name{new_lpme}
\alias{new_lpme}
\title{Create New LPME Object}
\usage{
new_lpme(
  embedding_map,
  msd,
  coefficients,
  times,
  initial_parameterization,
  optimal_parameterization,
  d,
  D,
  n_knots,
  lambda,
  gamma,
  coefficient_list,
  coefficient_functions,
  parameterization_list,
  smoothing_method,
  initialization_algorithm
)
}
\arguments{
\item{embedding_map}{The optimal embedding function.}

\item{msd}{Vector of Mean Squared Distance Values.}

\item{coefficients}{A list of matrices of spline coefficients.}

\item{times}{Numeric vector of the time values included in the dataset.}

\item{initial_parameterization}{A matrix of the initial parameters found for the data.}

\item{optimal_parameterization}{A matrix representing the best set of parameters.}

\item{d}{Integer value representing the intrinsic dimension of the manifold.}

\item{D}{Integer value representing the dimension of the input data.}

\item{n_knots}{The number of components identified.}

\item{lambda}{A vector of smoothing parameters for the pme functions at each time point.}

\item{gamma}{A vector of smoothing parameter values for the higher level spline model.}

\item{coefficient_list}{A list containing all matrices of coefficients.}

\item{coefficient_functions}{A list of functions that produce the PME coefficients at a given time point.}

\item{parameterization_list}{A list containing all matrices of parameters.}

\item{smoothing_method}{Choose between smoothing by "spline" or "gp".}

\item{initialization_algorithm}{The algorithm that was used to identify the initial parameter values.}
}
\value{
An object with class "lpme".
}
\description{
Create New LPME Object
}
