% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpme.R
\name{new_lpme}
\alias{new_lpme}
\title{Create New LPME Object}
\usage{
new_lpme(
  msd,
  sol_coef,
  times,
  r_init,
  r_fit,
  d,
  D,
  n_knots,
  lambda,
  gamma,
  sol_coef_list,
  r_fit_list
)
}
\arguments{
\item{msd}{Vector of Mean Squared Distance Values.}

\item{sol_coef}{A value.}

\item{times}{A value.}

\item{r_init}{A value.}

\item{r_fit}{A value.}

\item{d}{A value.}

\item{D}{A value.}

\item{n_knots}{A value.}

\item{lambda}{A value.}

\item{gamma}{A value.}

\item{sol_coef_list}{A value.}

\item{r_fit_list}{A value.}
}
\value{
An object with class "lpme".
}
\description{
Create New LPME Object
}
