% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{norm_euclidean}
\alias{norm_euclidean}
\title{Compute the Euclidean Norm of a Vector}
\usage{
norm_euclidean(x)
}
\arguments{
\item{x}{A numeric vector of values.}
}
\value{
A numeric value.
}
\description{
C++ implementation of the Euclidean norm calculations for a vector. This
improves computation time, but may introduce numerical instability.
}
\examples{
x <- 1:10
norm_euclidean(x)
}
