% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pme.R
\name{pme}
\alias{pme}
\title{Principal Manifold Estimation}
\usage{
pme(
  x_obs,
  d,
  initialization = NULL,
  N0 = 20 * D,
  tuning_para_seq = exp(-15:5),
  alpha = 0.05,
  max_comp = 100,
  epsilon = 0.05,
  max_iter = 100,
  SSD_ratio_threshold = 10,
  print_plots = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{x_obs}{Value}

\item{d}{Value}

\item{initialization}{Value}

\item{N0}{Value}

\item{tuning_para_seq}{Value}

\item{alpha}{Value}

\item{max_comp}{Value}

\item{epsilon}{Value}

\item{max_iter}{Value}

\item{SSD_ratio_threshold}{Value}

\item{print_plots}{Value}

\item{verbose}{Value}
}
\value{
An object of type pme.
}
\description{
This function still requires completed documentation
}
