% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{solve_spline}
\alias{solve_spline}
\title{Solve for Smoothing Spline Coefficients}
\usage{
solve_spline(E, t_val, X, w, d, D)
}
\arguments{
\item{E}{A matrix of values.}

\item{t_val}{A numeric matrix of input values.}

\item{X}{A numeric matrix of outputs.}

\item{w}{The smoothing parameter.}

\item{d}{The dimension of the input.}

\item{D}{The dimension of the output.}
}
\value{
A numeric matrix of coefficients.
}
\description{
Solve for Smoothing Spline Coefficients
}
